Optionally output compact constants instead of mirroring strings.
The function can now take a second parameter. If set to true (or any other truthy value), the module will output an object wherein the keys have the names of the strings that were passed in, but where the values are just contiguous integers starting at 0.

Also improved the tests a bit. Now using mocha and chai as test runner.